Issue #793 Allow to rename files, references and datasets - move test classes to packages
