+++ server/ngb-cli/src/test/java/com/epam/ngb/cli/manager/command/handler/http/dataset/DatasetListHandlerTest.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package com.epam.ngb.cli.manager.command.handler.http;
+package com.epam.ngb.cli.manager.command.handler.http.dataset;
 
 import com.epam.ngb.cli.AbstractCliTest;
 import com.epam.ngb.cli.TestDataProvider;
