+++ server/ngb-cli/src/test/java/com/epam/ngb/cli/manager/command/handler/http/reference/ReferenceDeletionHandlerTest.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package com.epam.ngb.cli.manager.command.handler.http;
+package com.epam.ngb.cli.manager.command.handler.http.reference;
 
 import com.epam.ngb.cli.AbstractCliTest;
 import com.epam.ngb.cli.TestHttpServer;
